refac: organizing details and request in files
