+++ Globoplay/Sources/Network/Requests/DiscoverRequest.swift
