+++ Globoplay/Sources/Network/Requests/SuggestionsRequest.swift
